move file (prompt)
