+++ prompt/prompt.go
@@ -1,4 +1,4 @@
-package cmd
+package prompt
 
 import (
 	"github.com/manifoldco/promptui"
